refactoring RemediationProcessor code
