+++ fcli-core/fcli-aviator-common/src/main/java/com/fortify/cli/aviator/fpr/remediation/SkipReason.java
@@ -1,4 +1,16 @@
-package com.fortify.cli.aviator.fpr.remediation.model;
+/*
+ * Copyright 2021-2026 Open Text.
+ *
+ * The only warranties for products and services of Open Text
+ * and its affiliates and licensors ("Open Text") are as may
+ * be set forth in the express warranty statements accompanying
+ * such products and services. Nothing herein should be construed
+ * as constituting an additional warranty. Open Text shall not be
+ * liable for technical or editorial errors or omissions contained
+ * herein. The information contained herein is subject to change
+ * without notice.
+ */
+package com.fortify.cli.aviator.fpr.remediation;
 
 public enum SkipReason {
     SOURCE_FILE_MISSING("Source file missing"),
@@ -19,7 +31,7 @@ public enum SkipReason {
     NO_CHANGES("No file changes found"),
     UNEXPECTED_ERROR("Unexpected remediation processing error");
 
-    public final String displayName;
+     final String displayName;
 
     SkipReason(String displayName) {
         this.displayName = displayName;
